fix: Clear finished work from WorkManager on launch
Owing to the persistent nature of WorkManager, when WorkManager is first created, we need to clear any finished work information so that no WorkInfo objects are present for the tagged "SaveToImageFileRequest" WorkRequest when the BlurActivity is launched. Doing so, it will not show the "See File" button as intended when the BlurActivity is launched. [fix #143](https://github.com/googlecodelabs/android-workmanager/issues/143)